+++ apps/console/src/extensions/components/application/api/application-roles.ts
@@ -32,7 +32,6 @@ import {
     AuthorizedAPIListItemInterface, 
     CreateRolePayloadInterface, 
     DescendantDataInterface, 
-    RolesV2ResponseInterface, 
     SharedApplicationAPIResponseInterface, 
     SharedApplicationDataInterface, 
     UpdateRolePayloadInterface
@@ -45,44 +44,6 @@ const httpClient: HttpClientInstance = AsgardeoSPAClient.getInstance()
     .httpRequest.bind(AsgardeoSPAClient.getInstance())
     .bind(AsgardeoSPAClient.getInstance());
 
-/**
- * Get the application roles by audience.
- *
- * @param audience - audience.
- * @param before - Before link.
- * @param after - After link.
- * @param limit - Limit.
- * 
- * @returns A promise containing the response.
- */
-export const getApplicationRolesByAudience = (
-    audience: string,
-    before: string,
-    after: string,
-    limit: number
-):Promise<RolesV2ResponseInterface> => {
-
-    const filter: string = `audience.type eq ${ audience.toLowerCase() }`;
-
-    const requestConfig: AxiosRequestConfig = {
-        method: HttpMethods.GET,
-        params: {
-            after,
-            before,
-            filter,
-            limit
-        },
-        url:  `${ store.getState().config.endpoints.rolesV2 }`
-    };
-
-    return httpClient(requestConfig)
-        .then((response: AxiosResponse) => {            
-            return Promise.resolve(response.data as RolesV2ResponseInterface);
-        })
-        .catch((error: AxiosError) => {
-            return Promise.reject(error);
-        });
-};
 
 /**
  * Get the application roles of the application.
