+++ apps/console/src/extensions/components/application/components/application-roles/application-roles.tsx
@@ -16,50 +16,59 @@
  * under the License.
  */
 
-import Autocomplete, {
-    AutocompleteRenderGetTagProps,
-    AutocompleteRenderInputParams
-} from "@mui/material/Autocomplete/Autocomplete";
-import Chip from "@oxygen-ui/react/Chip";
-import FormControlLabel from "@oxygen-ui/react/FormControlLabel";
-import FormGroup from "@oxygen-ui/react/FormGroup";
-import Radio from "@oxygen-ui/react/Radio";
-import TextField from "@oxygen-ui/react/TextField";
+import { AccessControlConstants, Show } from "@wso2is/access-control";
+import { IdentityAppsApiException } from "@wso2is/core/exceptions";
 import { AlertLevels, IdentifiableComponentInterface } from "@wso2is/core/models";
 import { addAlert } from "@wso2is/core/store";
+import { useTrigger } from "@wso2is/forms";
 import {
+    AnimatedAvatar,
+    AppAvatar,
     ConfirmationModal,
+    DataTable,
     DocumentationLink,
     EmphasizedSegment,
+    EmptyPlaceholder,
     Heading,
+    LinkButton,
+    ListLayout,
     PrimaryButton,
+    TableActionsInterface,
+    TableColumnInterface,
+    useConfirmationModalAlert,
     useDocumentation
 } from "@wso2is/react-components";
 import { AxiosError } from "axios";
 import React, {
     FunctionComponent,
-    HTMLAttributes,
+    MouseEvent,
     ReactElement,
+    ReactNode,
     SyntheticEvent,
+    useCallback,
     useEffect,
     useState
 } from "react";
 import { useTranslation } from "react-i18next";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { Dispatch } from "redux";
-import { Grid } from "semantic-ui-react";
-import { AutoCompleteRenderOption } from "./auto-complete-render-option";
-import { updateApplicationDetails } from "../../../../../features/applications/api";
-import { RoleAudienceTypes } from "../../../../../features/applications/constants/application-management";
-import { ApplicationInterface, BasicRoleInterface } from "../../../../../features/applications/models";
+import { Divider, DropdownProps, Grid, Header, Icon, PaginationProps, SemanticICONS } from "semantic-ui-react";
+import { CreateApplicationRoleWizard } from "./create-app-role-wizard";
+import { EditApplicationRole } from "./edit-app-role";
+import { ApplicationInterface } from "../../../../../features/applications/models";
 import {
+    AppState,
+    UIConstants,
+    getEmptyPlaceholderIllustrations,
     history
 } from "../../../../../features/core";
-import { getApplicationRolesByAudience } from "../../api/application-roles";
+import { OrganizationResponseInterface } from "../../../../../features/organizations/models";
+import { deleteRole, getApplicationRolesList, useSharedApplicationData } from "../../api/application-roles";
 import {
-    AssociatedRolesPatchObjectInterface,
-    RolesV2Interface,
-    RolesV2ResponseInterface
+    ApplicationRolesResponseInterface,
+    LinkInterface,
+    RoleListItemInterface,
+    SharedApplicationDataInterface
 } from "../../models/application-roles";
 
 interface ApplicationRolesSettingsInterface extends IdentifiableComponentInterface {
@@ -84,433 +93,569 @@ export const ApplicationRoles: FunctionComponent<ApplicationRolesSettingsInterfa
 
     const {
         application,
-        onUpdate,
         [ "data-componentid" ]: componentId
     } = props;
 
     const { t } = useTranslation();
     const dispatch: Dispatch<any> = useDispatch();
     const { getLink } = useDocumentation();
+    const [ alert, setAlert, alertComponent ] = useConfirmationModalAlert();
 
-    const [ isLoading, setIsLoading ] = useState<boolean>(false);
-    const [ isSubmitting, setIsSubmitting ] = useState<boolean>(false);
-    const [ showSwitchAudienceWarning, setShowSwitchAudienceWarning ] = useState<boolean>(false);
-    const [ roleAudience, setRoleAudience ] =
-        useState<RoleAudienceTypes>(application?.associatedRoles?.allowedAudience ?? RoleAudienceTypes.ORGANIZATION);
-    const [ tempRoleAudience, setTempRoleAudience ] =
-        useState<RoleAudienceTypes>(application?.associatedRoles?.allowedAudience ?? RoleAudienceTypes.ORGANIZATION);
-    
-    const [ roleList, setRoleList ] = useState<BasicRoleInterface[]>([]);
-    const [ selectedRoles, setSelectedRoles ] =
-        useState<BasicRoleInterface[]>(application?.associatedRoles?.roles ?? []);
-    const [ initialSelectedRoles, setInitialSelectedRoles ] =
-        useState<BasicRoleInterface[]>(application?.associatedRoles?.roles ?? []);
-    const [ activeOption, setActiveOption ] = useState<BasicRoleInterface>(undefined);
-    const [ removedRolesOptions, setRemovedRolesOptions ] = useState<BasicRoleInterface[]>(undefined);
+    const currentOrganization: OrganizationResponseInterface = useSelector(
+        (state: AppState) => state.organization.organization
+    );
+
+    const [ isLoading, setIsLoading ] = useState<boolean>(true);
+    const [ showWizard, setShowWizard ] = useState<boolean>(false);
+    const [ showEditModal, setShowEditModal ] = useState<boolean>(false);
+    const [ showDeleteConfirmationModal, setShowDeleteConfirmationModal ] = useState<boolean>(false);
+    const [ isDeleteSubmitting, setIsDeleteSubmitting ] =  useState<boolean>(false);
+
+    const [ roleList, setRoleList ] = useState<RoleListItemInterface[]>([]);
+    const [ roleListItem, setRoleListItem ] = useState<RoleListItemInterface>();
+    const [ applicationRoleResponse, setApplicationRoleResponse ] = useState<ApplicationRolesResponseInterface>(null);
+    const [ deletingRole, setDeletingRole ] = useState<RoleListItemInterface>(undefined);
+
+    const [ searchQuery, setSearchQuery ] = useState<string>("");
+    const [ listItemLimit, setListItemLimit ] = useState<number>(UIConstants.DEFAULT_RESOURCE_LIST_ITEM_LIMIT);
+    const [ after, setAfter ] = useState<string>("");
+    const [ before, setBefore ] = useState<string>("");
+    const [ isApplicationRoleNextPageAvailable, setIsApplicationRoleNextPageAvailable ] = useState<boolean>(undefined);
+    const [ isApplicationRoleNextPrevAvailable, setIsApplicationRolePrevPageAvailable ] = useState<boolean>(undefined);
+    const [ activePage, setActivePage ] = useState<number>(1);
+    const [ sharedApplications, setSharedApplications ] = useState<SharedApplicationDataInterface[]>([]);
+    const [ paginationReset, triggerResetPagination ] = useTrigger();
 
     const path: string[] = history.location.pathname.split("/");
     const appId: string = path[path.length - 1].split("#")[0];
+    const isSharedApplication: boolean = application?.advancedConfigurations?.fragment;
 
-    /**
-     * Fetch application roles on component load and audience switch.
-     */
-    useEffect(() => {        
-        getApplicationRoles();
-        if (roleAudience === application?.associatedRoles?.allowedAudience) {
-            setSelectedRoles(application?.associatedRoles?.roles);
-            setInitialSelectedRoles(application?.associatedRoles?.roles);
-        } else {
-            setSelectedRoles([]);
-            setInitialSelectedRoles([]);
+    const {
+        data: originalSharedApplicationData,
+        isLoading: isSharedApplicationDataFetchRequestLoading,
+        error: sharedApplicationDataFetchRequestError
+    } = useSharedApplicationData(appId, currentOrganization?.id);
+
+    useEffect(() => {                
+        if (originalSharedApplicationData instanceof IdentityAppsApiException 
+                || sharedApplicationDataFetchRequestError) {
+            handleRetrieveError();
+
+            return;
         }
-    }, [ roleAudience ]);
 
-    /**
-     * Set removed roles
-     */
+        if (!originalSharedApplicationData) {
+            return;
+        }
+
+        if (originalSharedApplicationData?.sharedApplications) {
+            setSharedApplications(originalSharedApplicationData?.sharedApplications);
+        }
+    }, [ originalSharedApplicationData ]);
+
     useEffect(() => {
-        if (initialSelectedRoles && selectedRoles) {
-            setRemovedRolesOptions(initialSelectedRoles?.filter((role: BasicRoleInterface) => {
-                return selectedRoles?.find(
-                    (selectedRole: BasicRoleInterface) => selectedRole.id === role.id) === undefined;
-            }));
+        let nextFound: boolean = false;
+        let prevFound: boolean = false;
+
+        applicationRoleResponse?.links?.forEach((link: LinkInterface) => {
+            if (link.rel === "after") {
+                const afterID: string = link.href.split("after=")[ 1 ];
+
+                setAfter(afterID);
+                setIsApplicationRoleNextPageAvailable(true);
+                nextFound = true;
+            }
+            if (link.rel === "before") {
+                const beforeID: string = link.href.split("before=")[ 1 ];
+
+                setBefore(beforeID);
+                setIsApplicationRolePrevPageAvailable(true);
+                prevFound = true;
+            }
+        });
+
+        if (!nextFound) {
+            setAfter("");
+            setIsApplicationRoleNextPageAvailable(false);
         }
-    }, [ initialSelectedRoles, selectedRoles ]);
+        if (!prevFound) {
+            setBefore("");
+            setIsApplicationRolePrevPageAvailable(false);
+        }
+    }, [ applicationRoleResponse ]);
+
+    useEffect(() => {
+        getApplicationRoles(appId, null, null, searchQuery? searchQuery: null , listItemLimit);
+    }, [ listItemLimit, searchQuery ]);
 
     /**
-     * Fetch application roles.
+     * Get application roles of the application.
+     *
+     * @param appId - Application ID.
+     * @param before - Before cursor link.
+     * @param after - After cursor link.
+     * @param filter - Filter query.
+     * @param limit - Limit.
      */
-    const getApplicationRoles = (): void => {
-        getApplicationRolesByAudience(roleAudience, null, null, null)
-            .then((response: RolesV2ResponseInterface) => {
-                const rolesArray: BasicRoleInterface[] = [];
+    const getApplicationRoles: (
+        appId: string,
+        before?: string,
+        after?: string,
+        filter?: string,
+        limit?: number
+    ) => void = useCallback(
+        (appId: string, before?: string, after?: string, filter?: string, limit?: number): void => {
+            getApplicationRolesList(appId, before, after, filter, limit)
+                .then((response: ApplicationRolesResponseInterface) => {
+                    setApplicationRoleResponse(response);
+                    setRoleList(response.roles);
+                }).catch((error: AxiosError) => {
+                    if (error?.response?.data?.description) {
+                        dispatch(addAlert({
+                            description: error?.response?.data?.description ??
+                                error?.response?.data?.detail ??
+                                t("extensions:develop.applications.edit.sections.roles.notifications." +
+                                    "fetchApplicationRoles.error.description"),
+                            level: AlertLevels.ERROR,
+                            message: error?.response?.data?.message ??
+                                t("extensions:develop.applications.edit.sections.roles.notifications." +
+                                    "fetchApplicationRoles.error.message")
+                        }));
 
-                response.Resources.forEach((role: RolesV2Interface) => {
-                    rolesArray.push({
-                        id: role.id,
-                        name: role.displayName
-                    });
-                });
-                
-                setRoleList(rolesArray);
-            }).catch((error: AxiosError) => {
-                if (error?.response?.data?.description) {
+                        return;
+                    }
                     dispatch(addAlert({
-                        description: error?.response?.data?.description ??
-                            error?.response?.data?.detail ??
-                            t("extensions:develop.applications.edit.sections.roles.notifications." +
-                                "fetchApplicationRoles.error.description"),
+                        description: t("extensions:develop.applications.edit.sections.roles.notifications." +
+                            "fetchApplicationRoles.genericError.description"),
                         level: AlertLevels.ERROR,
-                        message: error?.response?.data?.message ??
-                            t("extensions:develop.applications.edit.sections.roles.notifications." +
-                                "fetchApplicationRoles.error.message")
+                        message: t("extensions:develop.applications.edit.sections.roles.notifications." +
+                            "fetchApplicationRoles.genericError.message")
                     }));
 
-                    return;
-                }
+                    setApplicationRoleResponse({
+                        links: [],
+                        roles: []
+                    });
+                    setRoleList([]);
+                })
+                .finally(() => {
+                    setIsLoading(false);
+                });
+        }, [ getApplicationRolesList, setIsLoading ] );
+
+    /**
+     * Delete the selected application roles.
+     */
+    const deleteApplicationRole = (): void => {
+        setIsDeleteSubmitting(true);
+        deleteRole(appId, deletingRole.name)
+            .then(() => {
                 dispatch(addAlert({
                     description: t("extensions:develop.applications.edit.sections.roles.notifications." +
-                        "fetchApplicationRoles.genericError.description"),
+                        "deleteApplicationRole.success.description"),
+                    level: AlertLevels.SUCCESS,
+                    message: t("extensions:develop.applications.edit.sections.roles.notifications." +
+                        "deleteApplicationRole.success.message")
+                }));
+                onRoleUpdate();
+            }).catch(() => {
+                setAlert({
+                    description: t("extensions:develop.applications.edit.sections.roles.notifications." +
+                        "deleteApplicationRole.genericError.description"),
                     level: AlertLevels.ERROR,
                     message: t("extensions:develop.applications.edit.sections.roles.notifications." +
-                        "fetchApplicationRoles.genericError.message")
-                }));
-
-                setRoleList([]);
-            })
-            .finally(() => {
-                setIsLoading(false);
+                        "deleteApplicationRole.genericError.message")
+                });
+            }).finally(() => {
+                setIsDeleteSubmitting(false);
+                setShowDeleteConfirmationModal(false);
             });
+    };
+
+    /**
+     * Reset pagination to initial state.
+     */
+    const resetPagination: () => void = useCallback((): void => {
+        setActivePage(1);
+        triggerResetPagination();
+    }, [ setActivePage, triggerResetPagination ]);
+
+    /**
+     * Handles the search query clear action.
+     */
+    const handleSearchQueryClear: () => void = useCallback((): void => {
+        setSearchQuery("");
+        resetPagination();
+    }, [ setSearchQuery, resetPagination ]);
+
+    /**
+     * Handles the pagination change.
+     *
+     * @param event - Mouse event.
+     * @param data - Pagination component data.
+     */
+    const handlePaginationChange: (
+        event: MouseEvent<HTMLAnchorElement>,
+        data: PaginationProps
+    ) => void = useCallback((event: MouseEvent<HTMLAnchorElement>, data: PaginationProps): void => {
+        const newPage: number = parseInt(data?.activePage as string);
+
+        if (newPage > activePage) {
+            getApplicationRoles(appId, null, after, searchQuery, listItemLimit);
+        } else if (newPage < activePage) {
+            getApplicationRoles(appId, before, null, searchQuery, listItemLimit);
+        }
+        setActivePage(newPage);
+    }, [ activePage, searchQuery, listItemLimit, after, before ]);
+
+    /**
+     * Handles items per page change.
+     *
+     * @param event - Mouse event.
+     * @param data - Dropdown data.
+     */
+    const handleItemsPerPageDropdownChange: (
+        event: MouseEvent<HTMLAnchorElement>,
+        data: DropdownProps
+    ) => void = useCallback((event: MouseEvent<HTMLAnchorElement>, data: DropdownProps): void => {
+        setListItemLimit(data.value as number);
+        resetPagination();
+    }, [ setListItemLimit, resetPagination ]);
+
+    /**
+     * Handle the edit role action.
+     *
+     * @param role - Selected role.
+     */
+    const handleRoleEdit = (role: RoleListItemInterface) => {
+        setRoleListItem(role);
+        setShowEditModal(true);
+    };
+
+    /**
+     * Handle the delete role action.
+     *
+     * @param role - Selected role.
+     */
+    const handleRoleDelete = (role: RoleListItemInterface) => {
+        setDeletingRole(role);
+        setShowDeleteConfirmationModal(true);
     };
 
     /**
      * Triggers on role update.
      */
-    const updateRoles = (): void => {
-        setIsSubmitting(true);
-        const data: AssociatedRolesPatchObjectInterface = {
-            allowedAudience: roleAudience,
-            roles: selectedRoles ? selectedRoles.map((role: BasicRoleInterface) => {
-                return {
-                    id: role.id
-                };
-            }) : []
-        };
-
-        const updatedApplication: ApplicationInterface = {
-            associatedRoles: data,
-            id: appId,
-            name: application.name
-        };        
-
-        updateApplicationDetails(updatedApplication)
-            .then(() => {
-                dispatch(addAlert({
-                    description: t("console:develop.features.applications.notifications.updateApplication.success" +
-                        ".description"),
-                    level: AlertLevels.SUCCESS,
-                    message: t("console:develop.features.applications.notifications.updateApplication.success.message")
-                }));
-
-                onUpdate();
-            })
-            .catch((error: AxiosError) => {
-                if (error.response && error.response.data && error.response.data.description) {
-                    dispatch(addAlert({
-                        description: error.response.data.description,
-                        level: AlertLevels.ERROR,
-                        message: t("console:develop.features.applications.notifications.updateApplication.error" +
-                            ".message")
-                    }));
-
-                    return;
-                }
-
-                dispatch(addAlert({
-                    description: t("console:develop.features.applications.notifications.updateApplication" +
-                        ".genericError.description"),
-                    level: AlertLevels.ERROR,
-                    message: t("console:develop.features.applications.notifications.updateApplication.genericError" +
-                        ".message")
-                }));
-            })
-            .finally(() => {
-                setIsSubmitting(false);
-            });
+    const onRoleUpdate = (): void => {
+        getApplicationRoles(appId, null, null, null, null);
     };
 
     /**
-     * Handle the restore roles.
-     * 
-     * @param remainingRoles - remaining roles
+     * Displays the error banner when unable to fetch shared applications.
      */
-    const handleRestoreUsers = (remainingRoles: BasicRoleInterface[]) => {
-        const removedRoles: BasicRoleInterface[] = [];
+    const handleRetrieveError = (): void => {
+        dispatch(
+            addAlert({
+                description: t("extensions:console.applicationRoles.roleMapping.notifications.sharedApplication."+
+                    "error.description"),
+                level: AlertLevels.ERROR,
+                message: t("extensions:console.applicationRoles.roleMapping.notifications.sharedApplication."+
+                "error.message")
+            })
+        );
+    };
 
-        removedRolesOptions.forEach((user: BasicRoleInterface) => {
-            if (!remainingRoles?.find((newUser: BasicRoleInterface) => newUser.id === user.id)) {
-                removedRoles.push(user);
+    /**
+     * Resolves data table actions.
+     *
+     * @returns TableActionsInterface[]
+     */
+    const resolveTableActions = (): TableActionsInterface[] => {
+        return [
+            {
+                "data-componentid": `${ componentId }-item-edit-button`,
+                hidden: (): boolean => isSharedApplication,
+                icon: (): SemanticICONS => "pencil alternate",
+                onClick: (e: SyntheticEvent, role: RoleListItemInterface): void => handleRoleEdit(role),
+                popupText: (): string => t("common:edit"),
+                renderer: "semantic-icon"
+            },
+            {
+                "data-componentid": `${ componentId }-item-delete-button`,
+                hidden: (): boolean => isSharedApplication,
+                icon: (): SemanticICONS => "trash alternate",
+                onClick: (e: SyntheticEvent, role: RoleListItemInterface): void => handleRoleDelete(role),
+                popupText: (): string => t("common:delete"),
+                renderer: "semantic-icon"
+            },
+            {
+                "data-componentid": `${ componentId }-item-view-button`,
+                hidden: (): boolean => !isSharedApplication,
+                icon: (): SemanticICONS => "eye",
+                onClick: (e: SyntheticEvent, role: RoleListItemInterface): void => handleRoleEdit(role),
+                popupText: (): string => t("common:view"),
+                renderer: "semantic-icon"
             }
-        });
-
-        setSelectedRoles([
-            ...selectedRoles,
-            ...removedRoles
-        ]);
+        ];
     };
 
     /**
-     * Prompt the user to confirm the role audience switch.
-     * 
-     * @param selectedAudience - selected audience
+     * Resolves data table columns.
+     *
+     * @returns TableColumnInterface[]
      */
-    const promptAudienceSwitchWarning = (selectedAudience: RoleAudienceTypes): void => {
-        setTempRoleAudience(selectedAudience);
-        setShowSwitchAudienceWarning(true);
+    const resolveTableColumns = (): TableColumnInterface[] => {
+        return [
+            {
+                allowToggleVisibility: false,
+                dataIndex: "name",
+                id: "name",
+                key: "name",
+                render: (app: RoleListItemInterface): ReactNode => {
+                    return (
+                        <Header
+                            image
+                            as="h6"
+                            className="header-with-icon"
+                            data-componentid={ `${ componentId }-item-heading` }
+                        >
+                            <AppAvatar
+                                image={ (
+                                    <AnimatedAvatar
+                                        name={ app.name }
+                                        size="mini"
+                                        data-componentid={ `${ componentId }-item-image-inner` }
+                                    />
+                                ) }
+                                size="mini"
+                                spaced="right"
+                                data-componentid={ `${ componentId }-item-image` }
+                            />
+                            <Header.Content>
+                                { app.name }
+                            </Header.Content>
+                        </Header>
+                    );
+                },
+                title: t("extensions:develop.applications.edit.sections.roles.list.columns.name")
+            },
+            {
+                allowToggleVisibility: false,
+                dataIndex: "action",
+                id: "actions",
+                key: "actions",
+                textAlign: "right",
+                title: t("extensions:develop.applications.edit.sections.roles.list.columns.actions")
+            }
+        ];
+    };
+
+    /**
+     * Resolve the relevant placeholder.
+     *
+     * @returns React element.
+     */
+    const showPlaceholders = (): ReactElement => {
+        // When the search returns empty.
+        if (searchQuery && roleList.length === 0) {
+            return (
+                <EmptyPlaceholder
+                    action={ (
+                        <LinkButton onClick={ () => setSearchQuery("") }>
+                            { t("extensions:develop.applications.edit.sections.roles.placeHolders." +
+                                "emptySearchResults.action") }
+                        </LinkButton>
+                    ) }
+                    image={ getEmptyPlaceholderIllustrations().emptySearch }
+                    imageSize="tiny"
+                    title={ t("extensions:develop.applications.edit.sections.roles.placeHolders." +
+                        "emptySearchResults.title") }
+                    subtitle={ [
+                        t("extensions:develop.applications.edit.sections.roles.placeHolders." +
+                            "emptySearchResults.subtitles.0", { query: searchQuery })
+                    ] }
+                    data-componentid={ `${ componentId }-empty-search-placeholder-icon` }
+                />
+            );
+        }
+
+        if (roleList.length === 0) {
+            return (
+                <EmptyPlaceholder
+                    className={ "list-placeholder" }
+                    action={ !isSharedApplication &&
+                        (<Show when={ AccessControlConstants.APPLICATION_WRITE }>
+                            <PrimaryButton
+                                onClick={ () => { setShowWizard(true); } }>
+                                <Icon name="add" />
+                                { t("extensions:develop.applications.edit.sections.roles.placeHolders." +
+                                    "emptyList.action") }
+                            </PrimaryButton>
+                        </Show>)
+                    }
+                    image={ getEmptyPlaceholderIllustrations().newList }
+                    imageSize="tiny"
+                    title={ t("extensions:develop.applications.edit.sections.roles.placeHolders." +
+                        "emptyList.title") }
+                    subtitle={ [
+                        t("extensions:develop.applications.edit.sections.roles.placeHolders." +
+                            "emptyList.subtitles.0")
+                    ] }
+                    data-componentid={ `${ componentId }-empty-placeholder` }
+                />
+            );
+        }
+
+        return null;
     };
 
     return (
-        <>
-            <EmphasizedSegment
-                loading={ isLoading }
-                padded="very"
-                data-componentid={ componentId }
-            >
-                <Grid>
-                    <Grid.Row>
-                        <Grid.Column className="heading-wrapper" width={ 10 }>
-                            <Heading as="h4">
-                                { t("extensions:develop.applications.edit.sections.rolesV2.heading") }
-                            </Heading>
-                            <Heading subHeading ellipsis as="h6">
-                                { t("extensions:develop.applications.edit.sections.rolesV2.subHeading") }
-                                <DocumentationLink
-                                    link={ getLink("develop.applications.roles.learnMore") }
-                                >
-                                    { t("extensions:common.learnMore") }
-                                </DocumentationLink>
-                            </Heading>
-                        </Grid.Column>
-                    </Grid.Row>
-                    <Grid.Row>
-                        <Grid.Column width={ 10 }>
-                            <Heading as="h5">
-                                { t("extensions:develop.applications.edit.sections.rolesV2.roleAudience") }
-                            </Heading>
-                            <FormGroup className="additional-behaviors">
-                                <FormControlLabel
-                                    checked={ roleAudience === RoleAudienceTypes.ORGANIZATION }
-                                    control={ <Radio /> }
-                                    onChange={ () => promptAudienceSwitchWarning(RoleAudienceTypes.ORGANIZATION) }
-                                    label={ t("extensions:develop.applications.edit.sections.rolesV2.organization") }
-                                    data-componentid={ `${ componentId }-organization-audience-checkbox` }
-                                />
-                                <FormControlLabel
-                                    checked={ roleAudience === RoleAudienceTypes.APPLICATION }
-                                    control={ <Radio /> }
-                                    onChange={ () => promptAudienceSwitchWarning(RoleAudienceTypes.APPLICATION) }
-                                    label={ t("extensions:develop.applications.edit.sections.rolesV2.application") }
-                                    data-componentid={ `${ componentId }-application-audience-checkbox` }
-                                />
-                            </FormGroup>
-                        </Grid.Column>
-                    </Grid.Row>
-                    <Grid.Row className="p-0" verticalAlign="middle">
-                        <Grid.Column width={ 6 }>     
-                            <Heading as="h5">
-                                { t("extensions:develop.applications.edit.sections.rolesV2.assignedRoles") }
-                            </Heading>
-                        </Grid.Column>
-                        { /* {
-                            // Will be enabled once the feature is completed.
-                            roleAudience === RoleAudienceTypes.APPLICATION
-                                && (
-                                    <Grid.Column width={ 4 } textAlign="right">
-                                        <Button
-                                            startIcon={ <PlusIcon/> }
-                                            variant="text"
-                                        >
-                                            Create Role
-                                        </Button>
-                                    </Grid.Column>
-                                )
-                        } */ }
-                    </Grid.Row>
-                    <Grid.Row>
-                        <Grid.Column width={ 10 }>
-                            <Autocomplete
-                                multiple
-                                disableCloseOnSelect
-                                loading={ isLoading }
-                                options={ roleList }
-                                value={ selectedRoles ?? [] }
-                                getOptionLabel={ 
-                                    (role: BasicRoleInterface) => role.name
-                                }
-                                renderInput={ (params: AutocompleteRenderInputParams) => (
-                                    <TextField
-                                        { ...params }
-                                        placeholder={ t("extensions:develop.applications.edit.sections." +
-                                            "rolesV2.searchPlaceholder") }
-                                    />
-                                ) }
-                                onChange={ (event: SyntheticEvent, roles: BasicRoleInterface[]) => {
-                                    setSelectedRoles(roles); 
-                                } }
-                                isOptionEqualToValue={ 
-                                    (option: BasicRoleInterface, value: BasicRoleInterface) => 
-                                        option.id === value.id 
-                                }
-                                renderTags={ (
-                                    value: BasicRoleInterface[], 
-                                    getTagProps: AutocompleteRenderGetTagProps
-                                ) => value.map((option: BasicRoleInterface, index: number) => (
-                                    <Chip
-                                        { ...getTagProps({ index }) }
-                                        key={ index }
-                                        label={ option.name }
-                                        activeOption={ activeOption }
-                                        setActiveOption={ setActiveOption }
-                                        variant={
-                                            initialSelectedRoles?.find(
-                                                (role: BasicRoleInterface) => role.id === option.id
-                                            )
-                                                ? "solid"
-                                                : "outlined"
-                                        }
-                                    />
-                                )) }
-                                renderOption={ (
-                                    props: HTMLAttributes<HTMLLIElement>,
-                                    option: BasicRoleInterface, 
-                                    { selected }: { selected: boolean }
-                                ) => (
-                                    <AutoCompleteRenderOption
-                                        selected={ selected }
-                                        displayName={ option.name }
-                                        renderOptionProps={ props }
-                                    />
-                                ) }
-                            />
+        <EmphasizedSegment
+            loading={ isLoading || isSharedApplicationDataFetchRequestLoading }
+            padded="very"
+            data-componentid={ componentId }
+        >
+            <Grid>
+                <Grid.Row>
+                    <Grid.Column className="heading-wrapper" computer={ 10 }>
+                        <Heading as="h4">
+                            { t("extensions:develop.applications.edit.sections.roles.heading") }
+                        </Heading>
+                        <Heading subHeading ellipsis as="h6">
                             {
-                                removedRolesOptions?.length > 0
-                                    ? (
-                                        <Autocomplete
-                                            className="mt-3"
-                                            multiple
-                                            disableCloseOnSelect
-                                            loading={ isLoading }
-                                            options={ removedRolesOptions }
-                                            value={ removedRolesOptions }
-                                            getOptionLabel={ 
-                                                (role: BasicRoleInterface) => role.name
-                                            }
-                                            onChange={ (
-                                                event: SyntheticEvent,
-                                                remainingRoles: BasicRoleInterface[]
-                                            ) => {
-                                                handleRestoreUsers(remainingRoles);
-                                            } }
-                                            renderInput={ (params: AutocompleteRenderInputParams) => (
-                                                <TextField
-                                                    { ...params }
-                                                    label={ t("extensions:develop.applications.edit.sections." +
-                                                        "rolesV2.removedRoles") }
-                                                    placeholder={ t("extensions:develop.applications.edit.sections." +
-                                                        "rolesV2.searchPlaceholder") }
-                                                />
-                                            ) }
-                                            renderTags={ (
-                                                value: BasicRoleInterface[], 
-                                                getTagProps: AutocompleteRenderGetTagProps
-                                            ) => value.map((option: BasicRoleInterface, index: number) => (
-                                                <Chip
-                                                    { ...getTagProps({ index }) }
-                                                    key={ index }
-                                                    label={ option.name }
-                                                    option={ option }
-                                                    activeOption={ activeOption }
-                                                    setActiveOption={ setActiveOption }
-                                                    variant="outlined"
-                                                    onDelete={ () => {
-                                                        setSelectedRoles([
-                                                            ...selectedRoles,
-                                                            option
-                                                        ]);
-                                                    } }
-                                                />
-                                            ) ) }
-                                            renderOption={ (
-                                                props: HTMLAttributes<HTMLLIElement>, 
-                                                option: BasicRoleInterface
-                                            ) => (
-                                                <AutoCompleteRenderOption
-                                                    displayName={ option.name }
-                                                    renderOptionProps={ props }
-                                                />
-                                            ) }
-                                        />
-                                    ) : null
-                            } 
-                            
-                        </Grid.Column>
-                    </Grid.Row>
-                    <Grid.Row columns={ 1 } className="mt-5">
-                        <Grid.Column mobile={ 16 } tablet={ 16 } computer={ 16 }>
-                            <PrimaryButton
-                                size="small"
-                                loading={ isSubmitting }
-                                onClick={ () => updateRoles() }
-                                ariaLabel="Roles update button"
-                                data-componentid={ `${ componentId }-update-button` }
+                                isSharedApplication 
+                                    ? t("extensions:develop.applications.edit.sections.roles.subHeadingAlt")
+                                    : t("extensions:develop.applications.edit.sections.roles.subHeading")
+                            }
+                            <DocumentationLink
+                                link={ getLink("develop.applications.roles.learnMore") }
                             >
-                                { t("common:update") }
-                            </PrimaryButton>
-                        </Grid.Column>
-                    </Grid.Row>
-                </Grid>
-            </EmphasizedSegment>
-            <ConfirmationModal
-                onClose={ (): void => setShowSwitchAudienceWarning(false) }
-                type="negative"
-                open={ showSwitchAudienceWarning }
-                assertionHint={ t("extensions:develop.applications.edit.sections." +
-                    "rolesV2.switchRoleAudience.confirmationModal.assertionHint") }
-                assertionType="checkbox"
-                primaryAction={ t("common:confirm") }
-                secondaryAction={ t("common:cancel") }
-                onSecondaryActionClick={ (): void => {
-                    setShowSwitchAudienceWarning(false);
-                } }
-                onPrimaryActionClick={ (): void => {
-                    setRoleAudience(tempRoleAudience);
-                    setShowSwitchAudienceWarning(false);
-                } }
-                data-componentid={ `${ componentId }-switch-role-audience-confirmation-modal` }
-                closeOnDimmerClick={ false }
-            >
-                <ConfirmationModal.Header
-                    data-componentid={ `${ componentId }-switch-role-audience-confirmation-modal-header` }
-                >
-                    { t("extensions:develop.applications.edit.sections." +
-                        "rolesV2.switchRoleAudience.confirmationModal.header") }
-                </ConfirmationModal.Header>
-                <ConfirmationModal.Message
-                    attached
-                    negative
-                    data-componentid={ `${ componentId }-switch-role-audience-confirmation-modal-message` }
-                >
-                    { t("extensions:develop.applications.edit.sections." +
-                        "rolesV2.switchRoleAudience.confirmationModal.message") }
-                </ConfirmationModal.Message>
-                <ConfirmationModal.Content
-                    data-componentid={ `${ componentId }-switch-role-audience-confirmation-modal-content` }
-                >
-                    { t("extensions:develop.applications.edit.sections." +
-                        "rolesV2.switchRoleAudience.confirmationModal.content") }
-                </ConfirmationModal.Content>
-            </ConfirmationModal>
-        </>
+                                { t("extensions:common.learnMore") }
+                            </DocumentationLink>
+                        </Heading>
+                    </Grid.Column>
+                    <Grid.Column className="action-wrapper" computer={ 6 }>
+                        <div className="floated right action">
+                            {
+                                (roleList.length > 0) && !isSharedApplication && (
+                                    <PrimaryButton
+                                        data-componentid={ `${ componentId }-add-new-role-button` }
+                                        onClick={ () => setShowWizard(true) }
+                                    >
+                                        <Icon name="add" />
+                                        { t("extensions:develop.applications.edit.sections.roles.buttons.newRole") }
+                                    </PrimaryButton>
+                                )
+                            }
+                        </div>
+                    </Grid.Column>
+                </Grid.Row>
+            </Grid>
+            <Divider hidden />
+            <Grid>
+                <Grid.Row>
+                    <Grid.Column>
+                        <ListLayout
+                            currentListSize={ roleList.length }
+                            listItemLimit={ listItemLimit }
+                            onItemsPerPageDropdownChange={ handleItemsPerPageDropdownChange }
+                            onPageChange={ handlePaginationChange }
+                            showPagination={ true }
+                            showTopActionPanel={ false }
+                            showPaginationPageLimit={ false }
+                            totalPages={ 10 }
+                            totalListSize={ roleList.length }
+                            paginationOptions={ {
+                                disableNextButton: !isApplicationRoleNextPageAvailable,
+                                disablePreviousButton: !isApplicationRoleNextPrevAvailable
+                            } }
+                            resetPagination={ paginationReset }
+                            activePage={ activePage }
+                            data-componentid={ `${ componentId }-list-layout` }
+                        >
+                            <DataTable<RoleListItemInterface>
+                                className="application-roles-table"
+                                isLoading={ false }
+                                onSearchQueryClear={ handleSearchQueryClear }
+                                actions={ resolveTableActions() }
+                                columns={ resolveTableColumns() }
+                                data={ roleList }
+                                onRowClick={ (e: SyntheticEvent, role: RoleListItemInterface): void =>
+                                    handleRoleEdit(role) }
+                                placeholders={ showPlaceholders() }
+                                showHeader={ false }
+                                transparent={ !isLoading && (showPlaceholders() !== null) }
+                                data-componentid={ `${ componentId }-data-table` }
+                            />
+                        </ListLayout>
+                    </Grid.Column>
+                </Grid.Row>
+            </Grid>
+            <Divider hidden/>
+            {
+                showWizard && (
+                    <CreateApplicationRoleWizard
+                        data-componentid="create-app-role-wizard"
+                        onRoleUpdate={ onRoleUpdate }
+                        closeWizard={ () => setShowWizard(false) }
+                        appId={ appId }
+                        sharedApplications={ sharedApplications }
+                    />
+                )
+            }
+            <EditApplicationRole
+                data-componentid="edit-app-role-wizard"
+                onShowEditRoleModal={ setShowEditModal }
+                onRoleUpdate={ onRoleUpdate }
+                selectedRole={ roleListItem }
+                appId={ appId }
+                showEditRoleModal={ showEditModal }
+                isReadOnly={ isSharedApplication }
+            />
+            {
+                deletingRole && (
+                    <ConfirmationModal
+                        primaryActionLoading={ isDeleteSubmitting }
+                        onClose={ (): void => setShowDeleteConfirmationModal(false) }
+                        type="negative"
+                        open={ showDeleteConfirmationModal }
+                        assertionHint={ t("extensions:develop.applications.edit.sections.roles." +
+                                    "deleteRole.confirmationModal.assertionHint") }
+                        assertionType="checkbox"
+                        primaryAction={ t("common:confirm") }
+                        secondaryAction={ t("common:cancel") }
+                        onSecondaryActionClick={ (): void => {
+                            setShowDeleteConfirmationModal(false);
+                            setAlert(null);
+                        } }
+                        onPrimaryActionClick={ (): void => deleteApplicationRole() }
+                        data-componentid={ `${ componentId }-delete-confirmation-modal` }
+                        closeOnDimmerClick={ false }
+                    >
+                        <ConfirmationModal.Header
+                            data-componentid={ `${ componentId }-delete-confirmation-modal-header` }
+                        >
+                            { t("extensions:develop.applications.edit.sections.roles." +
+                                        "deleteRole.confirmationModal.header") }
+                        </ConfirmationModal.Header>
+                        <ConfirmationModal.Message
+                            attached
+                            negative
+                            data-componentid={ `${ componentId }-delete-confirmation-modal-message` }
+                        >
+                            { t("extensions:develop.applications.edit.sections.roles." +
+                                        "deleteRole.confirmationModal.message") }
+                        </ConfirmationModal.Message>
+                        <ConfirmationModal.Content
+                            data-componentid={ `${ componentId }-delete-confirmation-modal-content` }
+                        >
+                            <div className="modal-alert-wrapper"> { alert && alertComponent }</div>
+                            { t("extensions:develop.applications.edit.sections.roles." +
+                                        "deleteRole.confirmationModal.content") }
+                        </ConfirmationModal.Content>
+                    </ConfirmationModal>
+                )
+            }
+        </EmphasizedSegment>
     );
 };
