+++ apps/console/src/features/applications/constants/application-management.ts
@@ -445,8 +445,3 @@ export enum ShareWithOrgStatus {
     FALSE,
     UNDEFINED
 }
-
-export enum RoleAudienceTypes {
-    ORGANIZATION = "ORGANIZATION",
-    APPLICATION = "APPLICATION"
-}
