+++ apps/console/src/features/roles/models/roles.ts
@@ -19,6 +19,7 @@
 import { RolesInterface } from "@wso2is/core/models";
 import { CreateGroupMemberInterface } from "../../groups/models";
 import { SchemaAttributeValueInterface } from "../../users/models";
+import { RoleAudienceTypes } from "../constants/role-constants";
 
 /**
  * Interface to store data for create role api.
@@ -151,4 +152,61 @@ export interface CreateRoleStateInterface {
      * Basic details step form data.
      */
     [ CreateRoleStepsFormTypes.BASIC_DETAILS ]: CreateRoleFormData;
+}
+
+/**
+ * Interface for roles V2 data.
+ */
+export interface RolesV2Interface {
+    audience: {
+        display: string;
+        type: string;
+        value: string;
+    };
+    displayName: string;
+    id: string;
+    meta: {
+        location: string;
+    };
+}
+
+/**
+ *  Interface for roles V2 response data.
+ */
+export interface RolesV2ResponseInterface {
+    /**
+     * Number of results that match the listing operation.
+     */
+    totalResults?: number;
+    /**
+     * Index of the first element of the page, which will be equal to offset + 1.
+     */
+    startIndex?: number;
+    /**
+     * Schema related to the response.
+     */
+    schemas?: string[];
+    /**
+     * Number of elements in the returned page.
+     */
+    itemsPerPage?: number;
+    /**
+     * Set of roles.
+     */
+    Resources: RolesV2Interface[];
+}
+
+export interface AssociatedRolesPatchObjectInterface {
+    allowedAudience: RoleAudienceTypes;
+    roles: BasicRoleInterface[];
+}
+
+export interface AssociatedRolesInterface {
+    allowedAudience: RoleAudienceTypes;
+    roles: BasicRoleInterface[];
+}
+
+export interface BasicRoleInterface {
+    id: string;
+    name?: string;
 }
